requirements.txt
fastapi
uvicorn
youtube-transcript-api
langchain
langchain-google-genai
streamlit
gtts
requests
python-dotenv